Move DecaySwitch config option from property package to reaction package
